src/requirements.txt
dj-database-url==0.5.0
Django==2.0.4
django-appconf==1.0.2
django-bootstrap4==0.0.6
django-imagekit==4.0.2
gunicorn==19.7.1
pilkit==2.0
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
six==1.11.0
